Moving faketerm to internal.
