+++ container/traversal_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	"github.com/mum4k/termdash/internal/cell"
-	"github.com/mum4k/termdash/terminal/faketerm"
+	"github.com/mum4k/termdash/internal/faketerm"
 )
 
 func TestRoot(t *testing.T) {
