+++ internal/faketerm/diff.go
