+++ widgets/segmentdisplay/segmentdisplay_test.go
@@ -27,7 +27,7 @@ import (
 	"github.com/mum4k/termdash/internal/draw/segdisp/sixteen/testsixteen"
 	"github.com/mum4k/termdash/internal/terminalapi"
 	"github.com/mum4k/termdash/internal/widgetapi"
-	"github.com/mum4k/termdash/terminal/faketerm"
+	"github.com/mum4k/termdash/internal/faketerm"
 )
 
 // mustDrawChar draws the provided character in the area of the canvas or panics.
